+++ frontend/src/components/ProjectsComponents/CreateProjectComponents/AddedCoaches/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddedCoaches";
