+++ frontend/src/components/ProjectsComponents/CreateProjectComponents/InputFields/NumberOfStudents/NumberOfStudents.tsx
@@ -1,4 +1,4 @@
-import { Input } from "../styles";
+import { Input } from "../../styles";
 
 export default function NumberOfStudents({
     numberOfStudents,
