+++ frontend/src/components/ProjectsComponents/CreateProjectComponents/index.ts
@@ -1,0 +1,2 @@
+export { NameInput, NumberOfStudentsInput, CoachInput } from "./InputFields";
+export { default as AddedCoaches } from "./AddedCoaches";
