+++ frontend/src/views/projectViews/CreateProjectPage/CreateProjectPage.tsx
@@ -1,22 +1,15 @@
-import {
-    CreateProjectContainer,
-    Input,
-    AddButton,
-    RemoveButton,
-    CreateButton,
-    AddedCoach,
-} from "./styles";
+import { CreateProjectContainer, Input, AddButton, CreateButton } from "./styles";
 import { createProject } from "../../../utils/api/projects";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { GoBack } from "../ProjectDetailPage/styles";
 import { BiArrowBack } from "react-icons/bi";
-import { TiDeleteOutline } from "react-icons/ti";
 import {
     NameInput,
     NumberOfStudentsInput,
     CoachInput,
-} from "../../../components/ProjectsComponents/InputFields";
+    AddedCoaches,
+} from "../../../components/ProjectsComponents/CreateProjectComponents";
 
 export default function CreateProjectPage() {
     const [name, setName] = useState("");
@@ -24,6 +17,7 @@ export default function CreateProjectPage() {
     const [skills, setSkills] = useState([]);
     const [partners, setPartners] = useState([]);
 
+    // States for coaches
     const [coach, setCoach] = useState("");
     const [coaches, setCoaches] = useState<string[]>([]);
 
@@ -36,34 +30,22 @@ export default function CreateProjectPage() {
                 Cancel
             </GoBack>
             <h2>New Project</h2>
+
             <NameInput name={name} setName={setName} />
+
             <NumberOfStudentsInput
                 numberOfStudents={numberOfStudents}
                 setNumberOfStudents={setNumberOfStudents}
             />
+
             <CoachInput
                 coach={coach}
                 setCoach={setCoach}
                 coaches={coaches}
                 setCoaches={setCoaches}
             />
+            <AddedCoaches coaches={coaches} setCoaches={setCoaches} />
 
-            <div>
-                {coaches.map((element, _index) => (
-                    <AddedCoach key={_index}>
-                        {element}
-                        <RemoveButton
-                            onClick={() => {
-                                const newCoaches = [...coaches];
-                                newCoaches.splice(_index, 1);
-                                setCoaches(newCoaches);
-                            }}
-                        >
-                            <TiDeleteOutline size={"20px"} />
-                        </RemoveButton>
-                    </AddedCoach>
-                ))}
-            </div>
             <div>
                 <Input value={skills} onChange={e => setSkills([])} placeholder="Skill" />
                 <AddButton>Add skill</AddButton>
